Refactored code to accomodate new requirements
Moved commonly used methods from create_stack_dash and run_local_test to ui_tasks
